+++ java/src/tv/porst/swfretools/parser/structures/LineStyleParser.java
@@ -1,12 +1,13 @@
 package tv.porst.swfretools.parser.structures;
 
 import tv.porst.splib.io.BinaryParser;
+import tv.porst.splib.io.UINT8;
 
 public class LineStyleParser {
 
 	public static LineStyle parse(final BinaryParser parser) {
 
-		final int lineStyleType = parser.readUInt8();
+		final UINT8 lineStyleType = parser.readUInt8();
 		final RGB color = RGBParser.parse(parser);
 
 		return new LineStyle(lineStyleType, color);
